split tools and other tweaks
split tools into two pages to keep the api page focused on the REST API endpoint. added a note at the top of API pages clarifying API key is required. detailed the format of API key on the auth page. added aliases for assets and components to clean up imports.